fea: reuse home page
- move Home to the new route
- added layout to root
- update tailwindcss
- remove App
