+++ tailwind.config.cjs
@@ -2,7 +2,11 @@
 module.exports = {
   content: ["./src/**/*.{html,js,jsx,ts,tsx}"],
   theme: {
-    extend: {},
+    extend: {
+      fontFamily: {
+        serif: "Times New Roman",
+      },
+    },
   },
   plugins: [],
 };
